Picasso refact 2

+++ API/api.py
@@ -1,16 +1,25 @@
-from fastapi import APIRouter, Body
+from fastapi import APIRouter, Body, Header, HTTPException
 from pydantic import ValidationError
 from fastapi.responses import JSONResponse
 from API.urlSchema import AddUrlValidation, AddListUrlValidation, DeleteUrlValidation
-from utils.db import get_url, get_urls, delete_url, update_Url, create_Url
-from utils.gateWay import cache
+from utils.db import get_url, get_urls, delete_url, update_Url, create_Url, check_url_table_exists
+from gateWay import cache
+from dotenv import load_dotenv
+import os
 
 router = APIRouter(prefix='/url')
-
+load_dotenv()
 
 @router.post('/addUrl/')
-async def add_url(datas: AddListUrlValidation = Body()):
+async def add_url(datas: AddListUrlValidation = Body(),authorization: str = Header(None)):
+    correct_token = str(os.getenv("TOKEN"))
+    if authorization is None or authorization != correct_token:
+        raise HTTPException(status_code=401, detail="کاربر احراز هویت نشده است")
+    
+    await check_url_table_exists()
+    
     try:
+        
         for data in datas.data:
             # validations
             try:
@@ -19,14 +28,14 @@ async def add_url(datas: AddListUrlValidation = Body()):
                 return JSONResponse(content={"detail": str(e)}, status_code=400)
 
             # check if URL already exists
-            url = await get_url(data.signature)
+            url = await get_url(data.id)
 
             if url:
                 # update
-                await update_Url(data.path, data.signature, data.method, data.cache, url['id'])
+                await update_Url(data.id, data.path, data.signature, data.method, data.cache)
             else:
                 # create
-                await create_Url(data.path, data.signature, data.method, data.cache)
+                await create_Url(data.id, data.path, data.signature, data.method, data.cache)
 
         return JSONResponse(content={"detail": "URLs added successfully"}, status_code=200)
 
@@ -36,7 +45,13 @@ async def add_url(datas: AddListUrlValidation = Body()):
 
 
 @router.get('/getUrls/')
-async def get_urls_endpoint():
+async def get_urls_endpoint(authorization: str = Header(None)):
+    correct_token = str(os.getenv("TOKEN"))
+    if authorization is None or authorization != correct_token:
+        raise HTTPException(status_code=401, detail="کاربر احراز هویت نشده است")
+    
+    await check_url_table_exists()
+    
     try:
         # Assuming conn and cursor are available in the current scope
         urls_data = await get_urls()
@@ -47,7 +62,13 @@ async def get_urls_endpoint():
 
 
 @router.delete('/deleteUrl/')
-async def delete_url_endpoint(data: DeleteUrlValidation = Body()):
+async def delete_url_endpoint(data: DeleteUrlValidation = Body(),authorization: str = Header(None)):
+    correct_token = str(os.getenv("TOKEN"))
+    if authorization is None or authorization != correct_token:
+        raise HTTPException(status_code=401, detail="کاربر احراز هویت نشده است")
+
+    await check_url_table_exists()
+    
     try:
         # Assuming conn and cursor are available in the current scope
         await delete_url(data.id)
@@ -58,7 +79,13 @@ async def delete_url_endpoint(data: DeleteUrlValidation = Body()):
 
 
 @router.get('/clearCache/')
-async def clear_cache():
+async def clear_cache(authorization: str = Header(None)):
+    correct_token = str(os.getenv("TOKEN"))
+    if authorization is None or authorization != correct_token:
+        raise HTTPException(status_code=401, detail="کاربر احراز هویت نشده است")
+
+    await check_url_table_exists()
+        
     try:
         # Flush all items from the cache
         cache.flush_all()
@@ -72,5 +99,9 @@ async def clear_cache():
 
 
 @router.get("/health/")
-async def get_health():
+async def get_health(authorization: str = Header(None)):
+    correct_token = str(os.getenv("TOKEN"))
+    if authorization is None or authorization != correct_token:
+        raise HTTPException(status_code=401, detail="کاربر احراز هویت نشده است")
+
     return ({"status":200})
